管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -40,28 +40,32 @@ public class EmployeeAction extends ActionBase{
      * @throws IOException
      */
     public void index() throws ServletException, IOException{
-        //指定されたページ数の一覧画面に表示するデータを取得
-        int page = getPage();
-        List<EmployeeView> employees = service.getPerPage(page);
 
-        //すべての従業員データの件数を取得
-        long employeeCount = service.countAll();
+        //管理者かどうかのチェック
+        if(checkAdmin()) {
 
-        putRequestScope(AttributeConst.EMPLOYEES, employees);
-        putRequestScope(AttributeConst.EMP_COUNT, employeeCount);
-        putRequestScope(AttributeConst.PAGE, page);
-        putRequestScope(AttributeConst.MAX_ROW, JpaConst.ROW_PER_PAGE);
+            //指定されたページ数の一覧画面に表示するデータを取得
+            int page = getPage();
+            List<EmployeeView> employees = service.getPerPage(page);
 
-        //セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションから削除
-        String flush = getSessionScope(AttributeConst.FLUSH);
-        if (flush != null) {
-            putRequestScope(AttributeConst.FLUSH, flush);
-            removeSessionScope(AttributeConst.FLUSH);
+            //すべての従業員データの件数を取得
+            long employeeCount = service.countAll();
+
+            putRequestScope(AttributeConst.EMPLOYEES, employees);
+            putRequestScope(AttributeConst.EMP_COUNT, employeeCount);
+            putRequestScope(AttributeConst.PAGE, page);
+            putRequestScope(AttributeConst.MAX_ROW, JpaConst.ROW_PER_PAGE);
+
+            //セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションから削除
+            String flush = getSessionScope(AttributeConst.FLUSH);
+            if (flush != null) {
+                putRequestScope(AttributeConst.FLUSH, flush);
+                removeSessionScope(AttributeConst.FLUSH);
+            }
+
+            //一覧画面を表示
+            forward(ForwardConst.FW_EMP_INDEX);
         }
-
-        //一覧画面を表示
-        forward(ForwardConst.FW_EMP_INDEX);
-
     }
 
     /**
@@ -70,11 +74,16 @@ public class EmployeeAction extends ActionBase{
      * @throws IOException
      */
     public void entryNew() throws ServletException, IOException{
-        putRequestScope(AttributeConst.TOKEN, getTokenId());
-        putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());
 
-        //新規登録画面を表示
-        forward(ForwardConst.FW_EMP_NEW);
+        //管理者かどうかのチェック
+        if(checkAdmin()) {
+
+            putRequestScope(AttributeConst.TOKEN, getTokenId());
+            putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());
+
+            //新規登録画面を表示
+            forward(ForwardConst.FW_EMP_NEW);
+        }
     }
 
     /**
@@ -84,7 +93,7 @@ public class EmployeeAction extends ActionBase{
      */
     public void create() throws ServletException, IOException{
         //CSRF対策 tokenのチェック
-        if(checkToken()) {
+        if(checkToken() && checkAdmin()) {
             //パラメータの値を元に従業員情報のインスタンスを生成する
             EmployeeView ev = new EmployeeView(
                     null, //id
@@ -126,20 +135,24 @@ public class EmployeeAction extends ActionBase{
      * @throws IOException
      */
     public void show() throws ServletException, IOException{
-        //idを条件に従業員データを取得
-        EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-        if(ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+        //管理者かどうかのチェック
+        if(checkAdmin()) {
 
-            //データが取得できなかった、または論理削除されている場合はエラー画面を表示
-            forward(ForwardConst.FW_ERR_UNKNOWN);
-            return;
+            //idを条件に従業員データを取得
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+
+            if(ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+
+                //データが取得できなかった、または論理削除されている場合はエラー画面を表示
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
+
+            putRequestScope(AttributeConst.EMPLOYEE, ev);
+
+            forward(ForwardConst.FW_EMP_SHOW);
         }
-
-        putRequestScope(AttributeConst.EMPLOYEE, ev);
-
-        forward(ForwardConst.FW_EMP_SHOW);
-
     }
 
     /**
@@ -149,21 +162,24 @@ public class EmployeeAction extends ActionBase{
      */
     public void edit() throws ServletException, IOException{
 
-        //idを条件に従業員データを取得
-        EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+        //管理者かどうかのチェック
+        if(checkAdmin()) {
 
-        if(ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+            //idを条件に従業員データを取得
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-            //データが取得できなかった、または論理削除されている場合はエラー画面を表示
-            forward(ForwardConst.FW_ERR_UNKNOWN);
-            return;
+            if(ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+
+                //データが取得できなかった、または論理削除されている場合はエラー画面を表示
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
+
+            putRequestScope(AttributeConst.TOKEN, getTokenId());
+            putRequestScope(AttributeConst.EMPLOYEE, ev);
+
+            forward(ForwardConst.FW_EMP_EDIT);
         }
-
-        putRequestScope(AttributeConst.TOKEN, getTokenId());
-        putRequestScope(AttributeConst.EMPLOYEE, ev);
-
-        forward(ForwardConst.FW_EMP_EDIT);
-
     }
 
     /**
@@ -173,7 +189,7 @@ public class EmployeeAction extends ActionBase{
      */
     public void update() throws ServletException, IOException{
         //CSRF対策 tokenチェック
-        if(checkToken()) {
+        if(checkToken() && checkAdmin()) {
             //パラメータの値を元に従業員情報のインスタンスを作成
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -220,7 +236,7 @@ public class EmployeeAction extends ActionBase{
     public void destroy() throws ServletException, IOException{
 
         //CSRF対策 tokenチェック
-        if(checkToken()) {
+        if(checkToken() && checkAdmin()) {
 
             //idを条件に従業員データを論理削除する
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -232,4 +248,26 @@ public class EmployeeAction extends ActionBase{
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
         }
     }
+
+    /**
+     * ログイン中の従業員が管理者かどうかチェック。管理者でなければエラー画面を表示
+     * true：管理者
+     * @throws ServletException
+     * @throws IOException
+     */
+    private boolean checkAdmin() throws ServletException, IOException{
+
+        //セッションかログイン中の従業員情報を取得
+        EmployeeView ev = (EmployeeView)getSessionScope(AttributeConst.LOGIN_EMP);
+
+        //管理者でなければエラー画面を表示
+        if(ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+        }else {
+            return true;
+        }
+    }
+
 }
